app.py
```python
# Import necessary libraries and modules
from flask import Flask, render_template, request
from flask_socketio import SocketIO, emit
from dotenv import load_dotenv

# Import project-specific modules
import nlp
import ml
import format_standardization
import collaborative_writing
import feedback

# Load environment variables
load_dotenv()

# Initialize Flask app
app = Flask(__name__)
app.config['SECRET_KEY'] = 'secret!'
socketio = SocketIO(app)

@app.route('/')
def index():
    return render_template('index.html')

@socketio.on('script_input')
def handle_script_input(data):
    # Extract script data from client
    script = data['script']

    # Process script with AI modules
    script_outline = nlp.generate_outline(script)
    enhanced_dialogue = nlp.enhance_dialogue(script)
    standardized_script = format_standardization.standardize(script)
    narrative_feedback = feedback.analyze_script(script)
    collaborative_script = collaborative_writing.collaborate(script)

    # Emit processed script back to client
    emit('script_output', {
        'script_outline': script_outline,
        'enhanced_dialogue': enhanced_dialogue,
        'standardized_script': standardized_script,
        'narrative_feedback': narrative_feedback,
        'collaborative_script': collaborative_script
    })

if __name__ == '__main__':
    socketio.run(app)
```
