requirements.txt
# AI and Machine Learning
tensorflow>=2.4.1
torch>=1.8.1
transformers>=4.5.1

# Web Development
Flask>=1.1.2
flask-socketio>=5.0.1

# Natural Language Processing
nltk>=3.5
spacy>=3.0.6

# Data Handling
pandas>=1.2.4
numpy>=1.21.0

# Other
gunicorn>=20.1.0
python-dotenv>=0.17.1
